+++ frontend/src/components/templates/trim/TrimTemplate.tsx
@@ -1,4 +1,4 @@
-import { Header } from '@components/common/header/Header';
+import { Header } from '@components/common/Header/index';
 
 export function TrimTemplate() {
   return (
